Commented out 2 asserts which were getting triggered in Release Build.
The situations causing the asserts are handled properly in the code, thus the raised ASSERTS are for ultimately harmless reasons with regard to immediate integrity of the system.

The asserts sometimes occurred when a clip was allowed to play up to the end of the time line, where upon in a 3D situation an unexpected pixel format is being returned by adobe.